Bump to ASP.NET Core and C# 8
- Use of nullable reference types
- Switch to ASP.NET Core 3.0
- Switch to .NET Core 3.0
- Fixes #51
- Fixes #52
- Sample app configured to work with IIS Express

+++ src/FubarDev.WebDavServer/Engines/Local/CollectionTarget.cs
@@ -8,8 +8,6 @@ using System.Threading.Tasks;
 
 using FubarDev.WebDavServer.FileSystem;
 
-using JetBrains.Annotations;
-
 namespace FubarDev.WebDavServer.Engines.Local
 {
     /// <summary>
@@ -26,11 +24,11 @@ namespace FubarDev.WebDavServer.Engines.Local
         /// <param name="created">Was this collection created by the <see cref="RecursiveExecutionEngine{TCollection,TDocument,TMissing}"/>.</param>
         /// <param name="targetActions">The target actions implementation to use.</param>
         public CollectionTarget(
-            [NotNull] Uri destinationUrl,
-            [CanBeNull] CollectionTarget parent,
-            [NotNull] ICollection collection,
+            Uri destinationUrl,
+            CollectionTarget? parent,
+            ICollection collection,
             bool created,
-            [NotNull] ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
+            ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
             : base(targetActions, parent, destinationUrl, collection)
         {
             Collection = collection;
@@ -40,7 +38,6 @@ namespace FubarDev.WebDavServer.Engines.Local
         /// <summary>
         /// Gets the underlying collection.
         /// </summary>
-        [NotNull]
         public ICollection Collection { get; }
 
         /// <inheritdoc />
@@ -53,13 +50,12 @@ namespace FubarDev.WebDavServer.Engines.Local
         /// <param name="collection">The underlying collection.</param>
         /// <param name="targetActions">The target actions implementation to use.</param>
         /// <returns>The created collection target object.</returns>
-        [NotNull]
         public static CollectionTarget NewInstance(
-            [NotNull] Uri destinationUrl,
-            [NotNull] ICollection collection,
-            [NotNull] ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
+            Uri destinationUrl,
+            ICollection collection,
+            ITargetActions<CollectionTarget, DocumentTarget, MissingTarget> targetActions)
         {
-            CollectionTarget parentTarget;
+            CollectionTarget? parentTarget;
             if (collection.Parent != null)
             {
                 var collUrl = destinationUrl.GetParent();
@@ -77,6 +73,11 @@ namespace FubarDev.WebDavServer.Engines.Local
         /// <inheritdoc />
         public async Task<MissingTarget> DeleteAsync(CancellationToken cancellationToken)
         {
+            if (Parent == null)
+            {
+                throw new InvalidOperationException("Cannot delete entry, because the collection is unspecified.");
+            }
+
             var name = Collection.Name;
             await Collection.DeleteAsync(cancellationToken).ConfigureAwait(false);
             return new MissingTarget(DestinationUrl, name, Parent, TargetActions);
